backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
Pillow>=10.0
python-dateutil>=2.9
requests>=2.31
python-dotenv>=1.0
psycopg[binary]>=3.2